requirements.txt
PyYAML == 5.1.2
Click == 7